rename "Poll" Cog to "Polls"
