+++ cogs/components/poll/poll.py
